Removed File.ml and EventParser.ml and added some icons
The state representation has been vastly simplified which should
make debugging simpler as well. The idea to have a state object that
contained all of the constraints and graphics objects was a bad one,
and it has been removed.

EventParser.ml also contained too much hidden and immutable state,
so it has been removed in favor of a simpler but more transparent
system.